Fix vertex data

+++ game/shaders/VertexShader.glsl
@@ -3,7 +3,7 @@
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aNormal;
 layout (location = 2) in vec2 aUV;
-layout (location = 3) in float aData;
+layout (location = 3) in uint aData;
 
 uniform mat4 projection;
 uniform mat4 view;
@@ -18,8 +18,8 @@ out float ao;
 
 void main()
 {
-    texID = uint(aData) & 1023u;
-    ao = 1.0 - ((uint(aData) >> 10) & 3u) / 5.0;
+    texID = aData & 1023u;
+    ao = 1.0 -((aData >> 10) & 3u) / 5.0;
 
     uv = aUV;
     pos = aPos;
